feat: implement CB-2115 roadmap coherence
Pmat-Ticket: PMAT-722
